requirements.txt
gradio
pillow
imageio
imageio-ffmpeg
moviepy
numpy
openai>=1.51
python-dotenv
